pyproject.toml
[tool.poetry]
name = "gha-repo-manager"
version = "1.7.2"
description = "Manage your Github repo(s) settings and secrets using Github Actions and a yaml file"
authors = ["Andrew Herrington <andrew.the.techie@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "repo_manager"}]

[tool.poetry.scripts]
repo-manager = "repo_manager.main:main"


[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.1"
typing-extensions = "^4.7.1"
actions-toolkit = "^0.1.15"
pygithub = "^2.1.1"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
ruff = "^0.1.2"
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
bandit = "^1.7.8"
mypy = "^1.6.1"
pre-commit-hooks = "^4.5.0"
reorder-python-imports = "^3.12.0"
pytest-xdist = "^3.2.1"
pyupgrade = "^3.15.0"
pyflakes = "^3.1.0"
black = "^24.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.bandit]
exclude_dirs = ["tests", "noxfile.py", ".github/scripts"]

[tool.bandit.assert_used]
skips = ["repo_manager/github/branch_protections.py"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.mypy]
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.pytest.ini_options]
norecursedirs = ".github ci .git .idea"
addopts = "--cov=repo_manager --cov-report xml:.coverage.xml --cov-report=term-missing"
